+++ src/game/player/type.ts
@@ -1,1 +1,6 @@
+import {Choice} from "@core/elements/menu";
+
 export * from "@player/elements/type";
+export type Chosen = Choice & {
+    evaluated: string;
+};